Add awk script to prepare images for org-mode
The conversion from the original html files to org mode files is done using pandoc.
But pandoc ignores the alt attribute of the img tag. It would be nice for the alt
attribute to be used to generate a caption for the image in the resulting org mode
file. Fortunatelly, the original html file can be modified to replace all img tags
with the equivalent org mode image format plus captions. Only after this operation
is the html file to be converted to the org mode equivalent.