Nobody's gonna read the github version nor gonna modify it
The only reason to have two parts of the "curent continuu" chapter was that one part
is to big to be displayed in the github page. The hope was that the content would be
read by users on github. It was not the case. The chances are high the nobody will
want to modify the content from github as well. Having only one file instead of two
is easier to work with.